+++ packages/expo-sensors/plugin/src/withSensors.ts
@@ -1,9 +1,4 @@
-import {
-  AndroidConfig,
-  ConfigPlugin,
-  createRunOncePlugin,
-  withAndroidManifest,
-} from '@expo/config-plugins';
+import { ConfigPlugin, createRunOncePlugin } from '@expo/config-plugins';
 
 const pkg = require('expo-sensors/package.json');
 const MOTION_USAGE = 'Allow $(PRODUCT_NAME) to access your device motion';
@@ -17,41 +12,7 @@ const withSensors: ConfigPlugin<{ motionPermission?: string } | void> = (
   config.ios.infoPlist.NSMotionUsageDescription =
     motionPermission || config.ios.infoPlist.NSMotionUsageDescription || MOTION_USAGE;
 
-  return withAndroidFeature(config);
+  return config;
 };
-
-const withAndroidFeature: ConfigPlugin = config => {
-  return withAndroidManifest(config, config => {
-    config.modResults = setAndroidManifestFeatures(config.modResults);
-    return config;
-  });
-};
-
-/**
- * Add the following to the AndroidManifest.xml <uses-feature android:name="android.hardware.sensor.compass" android:required="true" />
- *
- * @param androidManifest
- */
-export function setAndroidManifestFeatures(
-  androidManifest: AndroidConfig.Manifest.AndroidManifest
-) {
-  if (!Array.isArray(androidManifest.manifest['uses-feature'])) {
-    androidManifest.manifest['uses-feature'] = [];
-  }
-  if (
-    !androidManifest.manifest['uses-feature'].find(
-      feature => feature.$['android:name'] === 'android.hardware.sensor.compass'
-    )
-  ) {
-    androidManifest.manifest['uses-feature'].push({
-      $: {
-        'android:name': 'android.hardware.sensor.compass',
-        'android:required': 'true',
-      },
-    });
-  }
-
-  return androidManifest;
-}
 
 export default createRunOncePlugin(withSensors, pkg.name, pkg.version);